+++ NEWS.md
@@ -1,12 +1,13 @@
 # memes 0.99.5
 * added `plot_sequence_heatmap` for making heatmaps of sequence lists
 * Added significantly more explanation to the ChIP-seq vignette
+* renamed `ame_plot_heatmap` -> `plot_ame_heatmap` for consistency 
+
 # memes 0.1.2
 * `runFimo()` `skip_matched_sequence` default is now `FALSE`. Set this to `TRUE` if fimo takes a long time to run, then use `add_sequence()` to add it back if needed.
 * `runTomTom()` `dist` default is now `ed` (changed from `pearson`).
 
 # memes 0.1.0
-
 * Removed `as_universalmotif_df()`, `as_universalmotif()`, and `update_motifs()`.
   * These functions are replaced by `universalmotif::to_df()`, `universalmotif::to_list()`, and `universalmotif::update_motifs()`
 * `runDreme` and `runTomTom` results are now returned in `universalmotif_df` format (behaves just like a data.frame)